added pop up navigation
