+++ package.json
@@ -10,6 +10,7 @@
     "axios": "^1.3.1",
     "jwt-decode": "^3.1.2",
     "react": "^18.2.0",
+    "react-burger-menu": "^3.0.9",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.0",
     "react-scripts": "5.0.1",
