+++ src/components/Navbar/Navbar.scss
@@ -1,19 +1,89 @@
 .nav {
   display: flex;
+  align-items: center;
   justify-content: space-between;
   padding: 1rem;
   background-color: white;
-  border-top: 1px solid $tertiary;
-  &-container {
-    display: flex;
-    &__icons {
-      width: 3rem;
-    }
+ 
+  &__menu{
+    width: 4rem;
+    height: 4rem;
   }
+
+  &__options{
+    justify-content: initial;
+  }
+
+  
 }
 
-.footer {
+
+/* Position and sizing of burger button */
+.bm-burger-button {
   position: fixed;
-  bottom: 0;
-  width: 100%;
+  width: 36px;
+  height: 30px;
+  right: 24px;
+  top: 26px;
+}
+
+/* Color/shape of burger icon bars */
+.bm-burger-bars {
+  background: #373a47;
+}
+
+/* Color/shape of burger icon bars on hover*/
+.bm-burger-bars-hover {
+  background: #a90000;
+}
+
+/* Position and sizing of clickable cross button */
+.bm-cross-button {
+  height: 24px;
+  width: 24px;
+}
+
+/* Color/shape of close button cross */
+.bm-cross {
+  background: #bdc3c7;
+}
+
+/*
+Sidebar wrapper styles
+Note: Beware of modifying this element as it can break the animations - you should not need to touch it in most cases
+*/
+.bm-menu-wrap {
+  position: fixed;
+  height: 100vh;
+  bottom: 0px;
+}
+
+/* General sidebar styles */
+.bm-menu {
+  background: $quad;
+  padding: 2.5em 1.5em 0;
+  font-size: 1.15em;
+}
+
+/* Morph shape necessary with bubble or elastic */
+.bm-morph-shape {
+  fill: #373a47;
+}
+
+/* Wrapper for item list */
+.bm-item-list {
+  color: #b8b7ad;
+  padding: 0.8em;
+}
+
+/* Individual item */
+.bm-item {
+  display: inline-block;
+  color: white;
+  margin: .5rem 0;
+}
+
+/* Styling of overlay */
+.bm-overlay {
+  background: rgba(0, 0, 0, 0.3);
 }
