+++ src/components/PageHeader/PageHeader.js
@@ -1,4 +1,5 @@
 import { Link } from "react-router-dom";
+import Nav from '../Navbar/Navbar'
 
 function PageHeader() {
   return (
@@ -7,6 +8,7 @@ function PageHeader() {
         <Link to="/">
           <h1 className="header__title">LegalLink</h1>
         </Link>
+        <Nav/>
       </header>
     </>
   );
