+++ src/components/PageHeader/PageHeader.scss
@@ -1,7 +1,9 @@
 .header {
+  display: flex;
+  justify-content: space-between;
+  background-color: white;
   &__title {
     color: $secondary;
-    background-color: white;
     margin: 0;
     padding: $mobile-pad;
   }
