+++ src/pages/Homepage/Homepage.scss
@@ -5,7 +5,7 @@
     display: block;
   }
   &__box {
-    background-color: #6f8565;
+    background-color: $tertiary;
     border-radius: 5px;
     padding: 2rem;
     display: flex;
