+++ src/styles/partials/_mixins.scss
@@ -19,3 +19,10 @@
   background-color: white;
   color: $secondary;
 }
+
+@mixin mobile-height {
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
